Move new dashboard to experimental implementation
